requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
discord.py==2.3.2
emoji==2.10.1
fonttools==4.50.0
frozenlist==1.4.1
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reloadium==1.4.0
requests==2.31.0
schema==0.7.5
six==1.16.0
urllib3==2.2.1
wavelink==3.2.0
yarl==1.9.4
